[Refactor] 파일명 jsx로 수정
